REFACTOR: fix typo mutli -> multi, foward -> forward
